+++ src/helpers/index.js
@@ -1,6 +1,6 @@
 const updateNotifier = require('./updateNotifier');
 const { configstore, resetConfigstore } = require('./configstore');
-const { getTheme, resetCustomTheme } = require('./getTheme');
+const { getTheme, resetCustomTheme } = require('./theme');
 const openUrls = require('./openUrls');
 
 module.exports = {