+++ src/index.js
@@ -1,5 +1,11 @@
 #!/usr/bin/env node
 
 const Community = require('./cli/Community');
+const parseArgs = require('minimist')(process.argv.slice(2));
 
-Community.start();
+(async () => {
+    await Community.start({
+        startCrawler: parseArgs._[0],
+        ...parseArgs,
+    });
+})();